Zomato-food-safety-advisor/requirements.txt
streamlit 

google.generativeai

python-dotenv